refactor: updateTemporaryPassword method to UserAuthService
